+++ astropy/units/tests/test_equivalencies.py
@@ -18,9 +18,13 @@ from ... import units as u
 
 def test_dimensionless_angles():
     rad1 = u.dimensionless_angles()
+    # test that the angles_dimensionless option allows one to change
+    # by any order in radian in the unit (#1161)
+    rad1 = u.angles_dimensionless()
     assert u.radian.to(1, equivalencies=rad1) == 1.
     assert u.deg.to(1, equivalencies=rad1) == u.deg.to(u.rad)
     assert u.steradian.to(1, equivalencies=rad1) == 1.
+    assert u.dimensionless_unscaled.to(u.steradian, equivalencies=rad1) == 1.
     # now quantities
     assert (1.*u.radian).to(1, equivalencies=rad1).value == 1.
     assert (1.*u.deg).to(1, equivalencies=rad1).value == u.deg.to(u.rad)
@@ -37,14 +41,6 @@ def test_dimensionless_angles():
 
     phase = MyRad1(1., u.cycle)
     assert phase.to(1).value == u.cycle.to(u.radian)
-
-
-def test_dimensionless_angles_context():
-    with u.set_enabled_equivalencies_context(u.angles_dimensionless()):
-        phase = u.Quantity(1., u.cycle)
-        assert_allclose(np.exp(1j*phase), 1.)
-        Omega = u.cycle / (1.*u.minute)
-        assert_allclose(np.exp(1j*Omega*60.*u.second), 1.)
 
 
 functions = [u.doppler_optical, u.doppler_radio, u.doppler_relativistic]
@@ -420,3 +416,25 @@ def test_brightness_temperature():
     np.testing.assert_almost_equal(
         1.0, tb.to(
             u.Jy, equivalencies=u.brightness_temperature(omega_B, nu)).value)
+
+
+def test_equivalency_context():
+    with u.set_enabled_equivalencies_context(u.angles_dimensionless()):
+        phase = u.Quantity(1., u.cycle)
+        assert_allclose(np.exp(1j*phase), 1.)
+        Omega = u.cycle / (1.*u.minute)
+        assert_allclose(np.exp(1j*Omega*60.*u.second), 1.)
+        # ensure we can turn off equivalencies even within the scope
+        with pytest.raises(u.UnitsError):
+            phase.to(1, equivalencies=[])
+
+    with u.set_enabled_equivalencies_context(u.spectral()):
+        u.GHz.to(u.cm)
+        eq_on = u.GHz.find_equivalent_units()
+        with pytest.raises(u.UnitsError):
+            u.GHz.to(u.cm, equivalencies=[])
+
+    # without equivalencies, we should find a smaller (sub)set
+    eq_off = u.GHz.find_equivalent_units()
+    assert all(eq in set(eq_on) for eq in eq_off)
+    assert set(eq_off) < set(eq_on)